Project: Move uploaded Files to their respective directories

+++ upload.java
@@ -3,12 +3,22 @@ import java.nio.file.Files;
 class upload{
 String path;
 File ok;
-upload(String path,String filename){
+upload(String path,String filename,String ext){
 try{
 File f = new File(path);
-File d = new File("/home/mrhacker5476/Desktop/"+filename);
+File dir=new File("/home/mrhacker5476/Desktop/"+ext);
+if(dir.exists()){
+File d = new File("/home/mrhacker5476/Desktop/"+ext+"/"+filename);
 Files.copy(f.toPath(),d.toPath());
 ok=d;
+}
+else{
+File dd = new File("/home/mrhacker5476/Desktop/"+ext);
+dd.mkdir();
+File d = new File("/home/mrhacker5476/Desktop/"+ext+"/"+filename);
+Files.copy(f.toPath(),d.toPath());
+ok=d;
+}
 }
 catch(Exception e){System.out.println(e.toString());}
 }
